Use standard inet_ntop().
git-svn-id: file:///home/yoann/SVN/trunk/prelude-manager@5753 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/include/server-generic.h
@@ -39,7 +39,7 @@
         prelude_msg_t *msg;          \
         int state;                   \
         char *addr;                  \
-        uint16_t port;               \
+        unsigned int port;           \
         uint64_t ident;              \
         char *permission_string;     \
         pthread_mutex_t mutex;       \
